fix: rebuild lib from src; correct login-shell quoting, chunk layout
- lib was hand-patched earlier; rebuild from src with tsdown so artifacts
  match the source exactly (quote-escaping typo in the hand-patch is gone)
- rolldown renamed the shared chunks (wsl-C5_mxGPM + wsl-credentials-*);
  the old wsl-CW3VPEIA chunk is no longer referenced and is removed
