+++ lib/fs.js
@@ -1,5 +1,4 @@
-import { a as joinUnc, c as parseWslUnc, l as windowsToMntPath, n as isAbsoluteLinuxPath, o as mntToWindowsPath } from "./paths-BDE1NVOv.js";
-import { n as defaultDistroSync } from "./wsl-CW3VPEIA.js";
+import { a as isAbsoluteLinuxPath, c as joinUnc, d as parseWslUnc, f as windowsToMntPath, l as mntToWindowsPath, n as defaultDistroSync } from "./wsl-C5_mxGPM.js";
 import z from "@deepseek-ai/schemastery";
 import { link, lstat, rename } from "node:fs/promises";
 import { FsError } from "@deepseek-ai/dsh-fs";
@@ -14,7 +13,7 @@ var WslFileSystem = class WslFileSystem extends LocalFileSystem {
 	static Config = z.object({
 		cwd: z.string(),
 		distro: z.string(),
-		diffBasisMaxBytes: z.number().default(10 * 1024 * 1024)
+		diffBasisMaxBytes: z.number().default(10485760)
 	});
 	distro;
 	constructor(ctx, config) {
@@ -101,6 +100,14 @@ var WslFileSystem = class WslFileSystem extends LocalFileSystem {
 		if (windowsToMntPath(base) !== null) return base;
 		throw new FsError(`wsl-fs: cwd "${base}" is not in the WSL execution world`, "FS_IO_ERROR");
 	}
+	/**
+	* Resolve the distribution an absolute Linux path opens inside. The chain:
+	* the caller cwd when it is a WSL UNC path, then the configured `distro`,
+	* then the host's default distribution from the Lxss registry (mirroring
+	* the shell executor's chain). Callers that pass no cwd — the
+	* `str_replace_editor` tool resolves with `{ signal }` only — still land
+	* on the same distro the session's bash runs in instead of failing.
+	*/
 	distroFor(cwd) {
 		const fromCwd = parseWslUnc(cwd ?? "");
 		if (fromCwd !== null) return fromCwd.distro;
