+++ lib/index.js
@@ -1,5 +1,5 @@
-import { a as joinUnc, c as parseWslUnc, n as isAbsoluteLinuxPath, o as mntToWindowsPath, r as isValidWslUsername, s as normalizeLinuxPath } from "./paths-BDE1NVOv.js";
-import { a as getWindowsWorkspace, c as registerWindowsWorkspace, i as canonicalWslUnc, l as setWorkspaceUsername, o as getWorkspaceUsername, r as listDistros, s as listWorkspaceKeys, t as defaultDistro } from "./wsl-CW3VPEIA.js";
+import { a as isAbsoluteLinuxPath, c as joinUnc, d as parseWslUnc, l as mntToWindowsPath, o as isValidWslUsername, r as listDistros, t as defaultDistro, u as normalizeLinuxPath } from "./wsl-C5_mxGPM.js";
+import { a as registerWindowsWorkspace, i as listWorkspaceKeys, n as getWindowsWorkspace, o as setWorkspaceUsername, r as getWorkspaceUsername, t as canonicalWslUnc } from "./wsl-credentials-BUP7ILY7.js";
 import z from "@deepseek-ai/schemastery";
 import { copyFileSync, existsSync, mkdirSync, readFileSync, readdirSync, rmSync, statSync, writeFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
@@ -20,7 +20,7 @@ import { homedir } from "node:os";
 * @module dsh-wsl-workspace/host/variants
 */
 /** Top-level rows that name the execution world and are replaced by the variant's own. */
-const WORLD_ROWS = new Set([
+const WORLD_ROWS = /* @__PURE__ */ new Set([
 	"tool-bash",
 	"tool-pwsh",
 	"tool-fs",
@@ -217,11 +217,11 @@ function variantDescription(presetId) {
 	const labels = MODE_DISPLAY_LABELS[presetId];
 	return `WSL execution world for ${labels === void 0 ? presetId : `${labels.en}（${labels.zh}）`}: bash and file tools run inside the WSL distribution.`;
 }
-const MAX_BODY_BYTES = 1024 * 1024;
+const MAX_BODY_BYTES = 1048576;
 /** Valid WSL distribution names: one path-safe segment (no separators, no dot-dirs). */
 const DISTRO_PATTERN = /^[A-Za-z0-9][A-Za-z0-9 ._-]*$/;
 /** The loopback hostnames the data route answers to (DNS-rebinding fence). */
-const LOOPBACK_HOSTNAMES = new Set([
+const LOOPBACK_HOSTNAMES = /* @__PURE__ */ new Set([
 	"localhost",
 	"127.0.0.1",
 	"::1",
@@ -318,7 +318,8 @@ function listWslDir(distro, linuxPath) {
 		};
 	}
 	const winPath = mntToWindowsPath(linuxPath);
-	const entries = readdirSync(winPath !== null ? winPath : joinUnc(distro, linuxPath), { withFileTypes: true }).slice(0, 1e3).map((dirent) => {
+	const readPath = winPath !== null ? winPath : joinUnc(distro, linuxPath);
+	const entries = readdirSync(readPath, { withFileTypes: true }).slice(0, 1e3).map((dirent) => {
 		const kind = dirent.isDirectory() ? "directory" : dirent.isFile() ? "file" : "other";
 		return {
 			name: dirent.name,
@@ -364,9 +365,11 @@ async function dispatch(method, params) {
 		}
 		case "registerWindows": {
 			const distro = requireDistro(params.distro);
-			const winPath = mntToWindowsPath(requireLinuxPath(params.linuxPath, "path"));
+			const linuxPath = requireLinuxPath(params.linuxPath, "path");
+			const winPath = mntToWindowsPath(linuxPath);
 			if (winPath === null) throw new Error("registerWindows requires a /mnt/<drive> Linux path");
-			registerWindowsWorkspace(winPath, distro, typeof params.username === "string" ? params.username : void 0);
+			const username = typeof params.username === "string" ? params.username : void 0;
+			registerWindowsWorkspace(winPath, distro, username);
 			return null;
 		}
 		case "listWorkspaces": return listWorkspaceKeys();
@@ -409,6 +412,15 @@ function localRowFiles(composition) {
 	return files;
 }
 /**
+* Preset data assets a composition's local plugin rows may read at mount
+* time, relative to the preset directory. The prefab-family presets load
+* `template.jsonl` (plus its sibling metadata and `templates/` directory)
+* from the seeder plugin's own directory; a variant missing them fails to
+* mount (ENOENT) and the session falls back to another preset.
+*/
+const PRESET_DATA_FILES = ["template.jsonl", "template.jsonl.meta.json"];
+const PRESET_DATA_DIRS = ["templates"];
+/**
 * Recursive regular-file directory copy for preset data assets. Deliberately
 * small: files and directories only, no symlink chasing — preset assets are
 * plain data the source preset already trusted.
@@ -427,7 +439,7 @@ function copyDirectorySync(src, dest) {
 * Missing assets are not an error — most presets carry none.
 */
 function copyPresetDataAssets(sourceDir, destDir, variantId) {
-	for (const dataFile of ["template.jsonl", "template.jsonl.meta.json"]) {
+	for (const dataFile of PRESET_DATA_FILES) {
 		const srcFile = join(sourceDir, dataFile);
 		if (!existsSync(srcFile)) continue;
 		try {
@@ -436,7 +448,7 @@ function copyPresetDataAssets(sourceDir, destDir, variantId) {
 			console.error(`dsh-wsl-workspace: failed to copy preset data asset ${dataFile} for ${variantId}: ${String(error instanceof Error ? error.message : error)}`);
 		}
 	}
-	for (const dataDir of ["templates"]) {
+	for (const dataDir of PRESET_DATA_DIRS) {
 		const srcDir = join(sourceDir, dataDir);
 		if (!existsSync(srcDir) || !statSync(srcDir).isDirectory()) continue;
 		try {
