+++ lib/shell.js
@@ -1,9 +1,9 @@
-import { a as joinUnc, c as parseWslUnc, i as isWindowsPathShaped, l as windowsToMntPath, r as isValidWslUsername } from "./paths-BDE1NVOv.js";
-import { n as defaultDistroSync, o as getWorkspaceUsername } from "./wsl-CW3VPEIA.js";
+import { c as joinUnc, d as parseWslUnc, f as windowsToMntPath, n as defaultDistroSync, o as isValidWslUsername, s as isWindowsPathShaped } from "./wsl-C5_mxGPM.js";
+import { r as getWorkspaceUsername } from "./wsl-credentials-BUP7ILY7.js";
 import z from "@deepseek-ai/schemastery";
 import { ShellExecutor } from "@deepseek-ai/dsh-shell";
 import { MAX_TIMER_DELAY_MS, clampTimeout, deadline, timeoutOf } from "@deepseek-ai/dsh-timeout";
-//#region \0@oxc-project+runtime@0.135.0/helpers/esm/usingCtx.js
+//#region \0@oxc-project+runtime@0.144.0/helpers/esm/usingCtx.js
 function _usingCtx() {
 	var r = "function" == typeof SuppressedError ? SuppressedError : function(r, e) {
 		var n = Error();
@@ -74,7 +74,7 @@ const ENV_OVERRIDES = {
 /** Default SIGTERM→SIGKILL grace period (matches `dsh-bash-local`). */
 const DEFAULT_GRACE_MS = 3e3;
 /** Default per-stream spill cap (matches `dsh-bash-local`). */
-const DEFAULT_MAX_SPILL_BYTES = 64 * 1024 * 1024;
+const DEFAULT_MAX_SPILL_BYTES = 67108864;
 /** Project a settled collect-mode reader into the final CollectedOutput shape. */
 function finalOutput(reader) {
 	const read = reader.readFrom(0);
@@ -191,7 +191,7 @@ var WslShellExecutor = class WslShellExecutor extends ShellExecutor {
 			username = this.resolveUser(spec, void 0);
 		}
 		const env = this.withWslEnv(spec);
-		const command = this.config.loginShell ? `cd '${linuxCwd.replace(/'/g, `'\''`)}' && ${spec.command}` : spec.command;
+		const command = this.config.loginShell ? `cd '${linuxCwd.replace(/'/g, `'\\''`)}' && ${spec.command}` : spec.command;
 		const argv = [
 			this.config.wslPath,
 			"-d",
